% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dl_wildfinder_data.R
\name{dl_wildfinder_data}
\alias{dl_wildfinder_data}
\title{Download wildfinder dataset}
\usage{
dl_wildfinder_data()
}
\value{
This function returns the path (\code{character}) to the downloaded file
(e.g. \code{data/wildfinder/wildfinder-ecoregions_list.csv}).
}
\description{
This function downloads the wildfinder dataset (text file) hosted on the
GitHub repository \url{https://github.com/rdatatoolbox/datarepo/}.
Three files \code{wildfinder-ecoregions_list.csv}, "wildfinder-ecoregions_species.csv",
and "wildfinder-mammals_list.csv" will be stored in
\verb{data/wildfinder/}. This folder will be created if required.
}
